requirements.txt
Flask
mysql-connector-python
PyJWT
pymongo